+++ settings.gradle
@@ -330,8 +330,7 @@ include 'testing:virtual-node-info-read-service-fake'
 include 'tools:plugins:db-config'
 include 'tools:plugins:initial-config'
 include 'tools:plugins:secret-config'
-include 'tools:tck:crypto-tck'
-include 'tools:tck:crypto-tck-impl'
+include 'tools:crypto-tck'
 
 // The following logic dictates if we use the composite build gradle functionality,
 // If this evaluates to true local source code will be used in place of binary dependencies.
