+++ tools/crypto-tck/src/main/java/net/corda/crypto/tck/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.crypto.tck;
+
+import org.osgi.annotation.bundle.Export;
