+++ tools/crypto-tck/src/main/kotlin/net/corda/crypto/tck/ComplianceTestType.kt
@@ -1,0 +1,6 @@
+package net.corda.crypto.tck
+
+enum class ComplianceTestType {
+    CRYPTO_SERVICE,
+    SESSION_INACTIVITY
+}
