+++ tools/crypto-tck/src/main/kotlin/net/corda/crypto/tck/CryptoTCK.kt
@@ -1,0 +1,5 @@
+package net.corda.crypto.tck
+
+interface CryptoTCK {
+    fun run(options: ExecutionOptions)
+}
