+++ tools/crypto-tck/src/main/kotlin/net/corda/crypto/tck/impl/compliance/SessionInactivityCompliance.kt
@@ -1,0 +1,10 @@
+package net.corda.crypto.tck.impl.compliance
+
+import org.junit.jupiter.api.Test
+
+class SessionInactivityCompliance {
+    @Test
+    fun `Blank session test`() {
+
+    }
+}
