+++ tools/crypto-tck/test.bndrun
@@ -1,0 +1,35 @@
+-tester: biz.aQute.tester.junit-platform
+-runfw: org.apache.felix.framework;
+-resolve.effective: resolve,active
+-runee: JavaSE-11
+-runtrace: true
+
+-runsystempackages: \
+    javax.xml.stream;version=1.0.0,\
+    javax.xml.stream.events;version=1.0.0,\
+    javax.xml.stream.util;version=1.0.0,\
+    sun.security.x509
+
+# Enable debugging.
+# -runjdb: 5055
+
+-runrequires: \
+    bnd.identity;id='net.corda.crypto-tck',\
+    bnd.identity;id='crypto-tck-tests',\
+    bnd.identity;id='junit-jupiter-engine',\
+    bnd.identity;id='junit-jupiter-params',\
+    bnd.identity;id='junit-platform-launcher',\
+    bnd.identity;id='junit-platform-reporting',\
+    bnd.identity;id='slf4j.simple'
+
+-runstartlevel: \
+    order=sortbynameversion,\
+    begin=-1
+
+-runproperties: \
+    org.osgi.framework.bsnversion=multiple,\
+    org.slf4j.simpleLogger.defaultLogLevel=info,\
+    org.slf4j.simpleLogger.showShortLogName=true,\
+    org.slf4j.simpleLogger.showThreadName=false,\
+    org.slf4j.simpleLogger.showDateTime=true,\
+    org.slf4j.simpleLogger.dateTimeFormat='yyyy-MM-dd HH:mm:ss:SSS Z'
