--- tools/tck/crypto-tck-impl/src/main/kotlin/net/corda/crypto/tck/impl/CryptoTCKImpl.kt
@@ -1,132 +1,0 @@
-package net.corda.crypto.tck.impl
-
-import net.corda.crypto.tck.impl.tests.CryptoServiceCompliance
-import net.corda.crypto.tck.impl.tests.SessionInactivityCompliance
-import net.corda.v5.base.util.contextLogger
-import net.corda.v5.crypto.tck.ComplianceTestType
-import net.corda.v5.crypto.tck.CryptoTCK
-import net.corda.v5.crypto.tck.ExecutionOptions
-import org.junit.jupiter.api.Assertions.assertTrue
-import org.junit.jupiter.engine.JupiterTestEngine
-import org.junit.platform.engine.discovery.DiscoverySelectors
-import org.junit.platform.launcher.Launcher
-import org.junit.platform.launcher.LauncherDiscoveryRequest
-import org.junit.platform.launcher.core.LauncherConfig
-import org.junit.platform.launcher.core.LauncherDiscoveryRequestBuilder
-import org.junit.platform.launcher.core.LauncherFactory
-import org.junit.platform.launcher.listeners.LoggingListener
-import org.junit.platform.launcher.listeners.SummaryGeneratingListener
-import org.junit.platform.launcher.listeners.TestExecutionSummary
-import org.junit.platform.reporting.legacy.xml.LegacyXmlReportGeneratingListener
-import org.opentest4j.TestAbortedException
-import org.osgi.service.component.annotations.Component
-import java.io.PrintWriter
-
-@Component(service = [CryptoTCK::class], immediate = true)
-class CryptoTCKImpl : CryptoTCK {
-    companion object {
-        private val logger = contextLogger()
-    }
-
-    private val version: String by lazy(LazyThreadSafetyMode.PUBLICATION) {
-        try {
-            CryptoTCKImpl::class.java.classLoader
-                .getResourceAsStream("META-INF/MANIFEST.MF")
-                .use {
-                    val manifest = java.util.jar.Manifest(it)
-                    val name = manifest.mainAttributes.getValue("Bundle-Name")
-                    val version = manifest.mainAttributes.getValue("Bundle-Version")
-                    arrayOf(
-                        name ?: "UNKNOWN",
-                        version ?: "UNKNOWN"
-                    )
-                }
-        } catch (e: Exception) {
-            emptyArray()
-        }.joinToString(" ")
-    }
-
-    override fun run(options: ExecutionOptions) = try {
-        val summaryListener = SummaryGeneratingListener()
-
-        val out = PrintWriter(System.out)
-
-        val request = buildTestRequest(options)
-
-        val launcher = buildTestLauncher(options, summaryListener, out)
-
-        printCaption(options)
-
-        launcher.execute(request)
-
-        summaryListener.printReport(out)
-
-        ExecutionOptionsExtension.unregister(options)
-
-        assertTrue(summaryListener.summary.isSuccess(), "TCK SUITE FAILED.")
-    } catch (e: Throwable) {
-        logger.error("TCK SUITE FAILED.", e)
-        throw e
-    }
-
-    private fun printCaption(options: ExecutionOptions) {
-        logger.info("EXECUTING COMPLIANCE TESTS: $version")
-        logger.info("${options::serviceName.name}=${options.serviceName}")
-        logger.info("${options::tests.name}=${options.tests.joinToString()}")
-        logger.info("${options::testResultsDirectory.name}=${options.testResultsDirectory}")
-        logger.info("--------------------------")
-    }
-
-    private fun buildTestRequest(options: ExecutionOptions): LauncherDiscoveryRequest {
-        val request = LauncherDiscoveryRequestBuilder.request()
-        if (options.tests.contains(ComplianceTestType.CRYPTO_SERVICE)) {
-            request.selectors(DiscoverySelectors.selectClass(CryptoServiceCompliance::class.java))
-        }
-        if (options.tests.contains(ComplianceTestType.SESSION_INACTIVITY)) {
-            request.selectors(DiscoverySelectors.selectClass(SessionInactivityCompliance::class.java))
-        }
-        ExecutionOptionsExtension.register(request, options)
-        return request.build()
-    }
-
-    private fun buildTestLauncher(
-        options: ExecutionOptions,
-        summaryListener: SummaryGeneratingListener,
-        out: PrintWriter
-    ): Launcher {
-        val launcherConfig = LauncherConfig.builder()
-            .enableTestEngineAutoRegistration(false)
-            .enableTestExecutionListenerAutoRegistration(false)
-            .addTestEngines(JupiterTestEngine())
-            .addTestExecutionListeners(summaryListener)
-            .addTestExecutionListeners(loggingListener())
-            .addTestExecutionListeners(LegacyXmlReportGeneratingListener(options.testResultsDirectory, out))
-            .build()
-        return LauncherFactory.create(launcherConfig)
-    }
-
-    private fun loggingListener(): LoggingListener =
-        LoggingListener.forBiConsumer { t, s ->
-            if (t != null) {
-                when (t) {
-                    is TestAbortedException -> logger.warn(s.get())
-                    is AssertionError -> logger.error(s.get())
-                    else -> logger.error(s.get(), t)
-                }
-            } else {
-                logger.info(s.get())
-            }
-        }
-
-    private fun SummaryGeneratingListener.printReport(out: PrintWriter) {
-        when (summary.isSuccess()) {
-            true -> out.println("${System.lineSeparator()}TESTS WERE SUCCESSFUL!")
-            else -> out.println("${System.lineSeparator()}TESTS HAVE FAILED!!!")
-        }
-        summary.printTo(out)
-        summary.printFailuresTo(out)
-    }
-
-    private fun TestExecutionSummary.isSuccess(): Boolean = containersFailedCount == 0L &&
-            testsFailedCount == 0L
-}
