--- tools/tck/crypto-tck-impl/src/main/kotlin/net/corda/crypto/tck/impl/ExecutionOptionsExtension.kt
@@ -1,38 +1,0 @@
-package net.corda.crypto.tck.impl
-
-import net.corda.v5.crypto.tck.ExecutionOptions
-import org.junit.jupiter.api.extension.ExtensionContext
-import org.junit.jupiter.api.extension.ParameterContext
-import org.junit.jupiter.api.extension.ParameterResolver
-import org.junit.platform.launcher.core.LauncherDiscoveryRequestBuilder
-import java.util.UUID
-import java.util.concurrent.ConcurrentHashMap
-
-class ExecutionOptionsExtension : ParameterResolver {
-    companion object {
-        private const val CONFIGURATION_PARAMETER_KEY = "tck.executionOptions"
-
-        private val map = ConcurrentHashMap<String, ExecutionOptions>()
-
-        fun register(requestBuilder: LauncherDiscoveryRequestBuilder, options: ExecutionOptions) {
-            val id = UUID.randomUUID().toString()
-            this.map[id] = options
-            requestBuilder.configurationParameter(CONFIGURATION_PARAMETER_KEY, id)
-        }
-
-        fun unregister(options: ExecutionOptions) {
-            val found = map.filterValues { it == options }.entries
-            found.forEach {
-                map.remove(it.key)
-            }
-        }
-    }
-
-    override fun supportsParameter(parameterContext: ParameterContext, extensionContext: ExtensionContext): Boolean =
-        parameterContext.parameter.type.isAssignableFrom(ExecutionOptions::class.java)
-
-    override fun resolveParameter(parameterContext: ParameterContext, extensionContext: ExtensionContext): Any {
-        val id = extensionContext.getConfigurationParameter(CONFIGURATION_PARAMETER_KEY)
-        return map.getValue(id.get())
-    }
-}
