--- tools/tck/crypto-tck-impl/src/main/kotlin/net/corda/crypto/tck/impl/tests/SessionInactivityCompliance.kt
@@ -1,10 +1,0 @@
-package net.corda.crypto.tck.impl.tests
-
-import org.junit.jupiter.api.Test
-
-class SessionInactivityCompliance {
-    @Test
-    fun `Blank session test`() {
-
-    }
-}
