--- tools/tck/crypto-tck/build.gradle
@@ -1,17 +1,0 @@
-plugins {
-    id 'corda.common-publishing'
-    id 'corda.common-library'
-}
-
-description 'Crypto TCK public API'
-
-dependencies {
-    compileOnly 'org.osgi:osgi.annotation'
-
-    implementation 'net.corda.kotlin:kotlin-stdlib-jdk8-osgi'
-    implementation platform("net.corda:corda-api:$cordaApiVersion")
-
-    api "net.corda:corda-cipher-suite"
-}
-
-
