--- tools/tck/crypto-tck/src/main/java/net/corda/v5/crypto/tck/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.v5.crypto.tck;
-
-import org.osgi.annotation.bundle.Export;
