--- tools/tck/crypto-tck/src/main/kotlin/net/corda/v5/crypto/tck/ComplianceTestType.kt
@@ -1,6 +1,0 @@
-package net.corda.v5.crypto.tck
-
-enum class ComplianceTestType {
-    CRYPTO_SERVICE,
-    SESSION_INACTIVITY
-}
