--- tools/tck/crypto-tck/src/main/kotlin/net/corda/v5/crypto/tck/CryptoTCK.kt
@@ -1,5 +1,0 @@
-package net.corda.v5.crypto.tck
-
-interface CryptoTCK {
-    fun run(options: ExecutionOptions)
-}
